Add testing entry point for main runner
Using injected parameters, rather than from CLI
